trusted-firmware-m: partition: Add an example partition
Add an example partition which provides a secure SHA-256 hash service
that computes the SHA-256 hash of a given message and returns the hash.
In addition, it also contains a templated example which can be used as
a starting point for developing secure services. The templated example
service shows how to start a platform timer and implement a timer IRQ
handler.

Signed-off-by: Devaraj Ranganna <devaraj.ranganna@linaro.org>