Build: Allow customized manifest lists
The tfm_manifest_list.yaml is used by default and only extra
manifest lists are allowed for customization.

This patch makes the default manifest list replacable as well.

This can avoid issues that the extra manifest and the default
manifest may have conflict Partitions or there could be conflicts
within the default manifest list on some specific builds.

Note: The Partition configuration switches defined in "conditional"
are still required to set to OFF to disable building Partitions even
if they are not in manifest lists.

This patch also fixes the wrong directory in config/set_config.cmake
which happens to be correct when building.

Change-Id: Id701e174d511c8fe34d62e8a3ff4813460aab2e9
Signed-off-by: Kevin Peng <kevin.peng@arm.com>