Platform: Add a customized manifest list for Musca-B1
When the Secure Enclave is enabled, the host core does not provide
PSA RoT Services and Protected Storage services. Instead a Proxy
Partition is provided to forward service requests to the SE.

The manifest list does not need to include those Partitions. Having
those Partitions causes potential conflicts with the Proxy Partition
because they provide same services and use the same manifest items.

This patch adds a customized manifest list for Musca-B1 with Secure
Encalve and removes the Proxy Partition in the common manifest list.

Change-Id: Ib83ee415b7b3fc8a53eb59e2de42964f8c5fe326
Signed-off-by: Kevin Peng <kevin.peng@arm.com>