Tools: Fixed slashes mismatch in manifest parsing tool
When generating manifest files on Windows and the output option (-o) is
specified, there are different types of slashes in the include paths of
generated files (the part of specified folder uses Windows slashes (\)
whereas the rest of path uses Unix slashes (/)). This patch unified it
to use only Unix slashes.

Also it was deleted converting to absolute paths functionality when
specifying output folder, now it uses relative paths as it was
originally. This allows to update files in their locations even outside
the root folder of the directory.

Signed-off-by: MartinaHanusovaNXP <martina.hanusova@nxp.com>
Change-Id: I9469d658afbce001ce45abe2e61ccc2dc4b5a02d