Proxy: Don't include platform_multicore.h
There are two versions of platform_multicore.h in the TF-M
source, both under platform/ext/target/... Partition code
should not be including header files from the platform directory
except where such header files are part of the platform API
(such as region_defs.h).

It seems in this case that the code doesn't actually need the #include,
so just remove it.

Signed-off-by: Chris Brand <chris.brand@cypress.com>
Change-Id: Ib651587fe4ec08e3850fab56f27e111c81b8964b