Partitions: Add psa_wait when wake up in Idle Partition
There could be the following case that SPE would stuck in Idle
Partition:
 - The NSPE preemptes the Idle Partition
 - A Secure Interrupt then preempts the NSPE
 - The interrupt handling triggers scheduling.
 - But the NSPE was preempted, scheduling is skipped to avoid context
   mismatch between NSPE and SPE
 - Execution goes back from NSPE to SPE

At this point, some events from the resumed Partition to trigger the
scheduler again is expected, for example a psa_wait or psa_reply call.
But the Idle Partition does nothing.

So this patch adds a dummy psa_wait to fix this issue, using PSA_POLL
to avoid the Idle Partition being set to "BLOCK" state.

Change-Id: I69e79b25ba88eea40d18b1c7c29690bc1cadf242
Signed-off-by: Kevin Peng <kevin.peng@arm.com>