platform: nordic: Fix build error in configurations without BL2
This is a follow-up to commit 32a7431436b7ee7110880f128e0519900aede767.

Do not reserve trailer space when building without BL2, as in such case
the veneers section is placed at the end of the image (position 3 in
the linker script) and this reserved space would cause the FLASH region
to be overflowed for bigger images - in the nRF IDAU, the NSC section
containing veneers can be placed only at the very end of the selected
secure flash region in hardware, so the reserved space would be cut out
from this section.

Without this change, building fails e.g. for the following parameters:
-DTEST_S=ON -DTEST_NS=ON -DBL2=OFF

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
Change-Id: If033d96ef71bce628335f2b9f819ccb83d741df8