HAL: Migrate legacy interrupt API to new HAL API
This patch migrates the legacy interrupt APIs to the current HAL.
Only three of them are migrated, they are:
 - tfm_spm_hal_enable_irq
 - tfm_spm_hal_disable_irq
 - tfm_spm_hal_clear_pending_irq
The implementations are the same, and are same across most platforms.
So a common implementation bases on NVIC is provided.
Platforms can choose to use it directly or have a dedicated version.

The old ones are kept for Library Model.

Others APIs will be replaced by interrupt binding API.

Change-Id: I5842db8cd983bc6323aa24329822b13e3fbc9691
Signed-off-by: Kevin Peng <kevin.peng@arm.com>