Build: Use IPC Model to build by default
The Library Model (TFM_PSA_API=OFF) is not being developed anymore.
It does not support for new FF-M features.

New comers to TF-M should be encouraged to start with the IPC model
to have the better experiences.

Library Model might be replaced by SFN Model in the future.

This patch adds a new TFM_LIB_MODEL configuration and set to OFF by
default to build IPC model by default.

The TFM_PSA_API is not intended for users to choose between library
and IPC model anymore.
To build Library model, please set TFM_LIB_MODEL to ON.

Note:
The test repo version is updated as well for the dependency changes.

Change-Id: I43ec92cd46dbc5a26e3638269cfb6267e24352bd
Signed-off-by: Kevin Peng <kevin.peng@arm.com>