corstone1000: support secure debug
The commit integrates trusted-firmware-m/corstone1000 target
of psa-adac. SDC-600 COM PORT is the communication medium
between host and the target. The driver for the SDC-600 and
Corstone1000 platform integration to psa-adac's platform
HAL is part of psa-adac git repository. This commit just
enables the secure debug workflow and integrates the
api of cc312 to set the debug permission vectors.

The commit also configures the firewall to allow access to
APBCOM port. It also increases the BL2 stack size as previous
stack size was not enough to run the psa-adac secure debug
software.

Change-Id: I7d8c3cac3002874913e3eb6242150691ae091cda
Signed-off-by: Satish Kumar <satish.kumar01@arm.com>