Revert "SPM: Correct the PendSV priority value"
This reverts commit 91d9f7403cb476539b74cc06de82483977daf3a0.

The patch didn't take the following case into account:
The NSPE masks NS interrupts by setting the PRIMASK_NS to 1 before
requesting secure services.

Although the NS execution has the priority of 0x80, which is same as
Secure PendSV, with the effect of AIRCR.PRIS.
The Secure PendSV still does not have enough priority to preempt the
NSPE - It needs to have a higher priority, being equal is not enough.

Change-Id: I6757adc81d4bc3decc656afea8e6fb3a91bad136
Signed-off-by: Kevin Peng <kevin.peng@arm.com>