Platform: Duplicate tfm_spm_hal_has_access_to_region()
This patch duplicates the tfm_spm_hal_has_access_to_region() for IPC
Model.

In IPC model, the whole memory access check function is HAL API.
There is no need to call another HAL API within that API. So the
API is renamed as an internal function in tfm_hal_isolation.c.

Note:
Only the Nordic platforms has the implementation of this API.

Change-Id: I956e3733f6679a9ae5c27861bb706fc5aea01760
Signed-off-by: Kevin Peng <kevin.peng@arm.com>