ITS: Add metadata XOR checksum to validate metadata
ITS file system operates the files based on the metadata. Currently
only the hardcoded fs_version and the swap count is checked when
validating the metadata. It is not enough to check against the invalid
data in the flash. So the XOR value based on all the metadata(block metadata
and file metadata) is added into the meta_block_header. The XOR value
check is added when validating the metadata.

Signed-off-by: Sherry Zhnag <sherry.zhang2@arm.com>
Change-Id: Ief9da88bb0ef81b3b7610c0430e139a3fa8e14af