Boot: Add --measured-boot-record option to wrapper.py
When the --measured-boot-record flag is enabled the Python script adds
the CBOR encoded measured boot record to the image manifest (to the TLV
area). Previously it was the default behavior, however this flag's
default value is false, therefore from now on it has to be used to
achieve the same result.

Change-Id: Ia689b27dddc83ec5a2be4f56782ca9e355852025
Signed-off-by: David Vincze <david.vincze@arm.com>