PsoC64: Fix missed removal of switch to PSP in ARMClang startup.s
Commit 300c68da11fda3669d7b80b68dcd67b0b2082cbc (SPM: Use Main Stack
for initialization) changes all platforms to use main stack instead of
processor stack for initialization by removing the switch to processor
stack. However, the removal of the PSP switch was missed in the ARMClang
startup.s for PsoC64 leading to boot failure.

This commit removes the PSP switch in PSoC64 ARMClang startup.s in the
same manner as was done for other platforms.

Signed-off-by: Alexander Moore <alexander.moore@cypress.com>
Signed-off-by: Chris Brand <chris.brand@cypress.com>
Change-Id: I37ecc8d35460208501d2033fee7eb9ce26c8be90