Tool: Calculate total size of stacks for SFN backend
Under the SFN backend in isolation level 1, partitions do not need
dedicated stacks since they can work on the caller stack. Let the
tool calculate the sum of stacks and provide for further usage.
NS Agent can allocate stack with this sum since it is the initial
caller of all access.

Change-Id: Iee4ae996fc43cf1fc7f4c1d468372ae5c194bb4f
Signed-off-by: Mingyang Sun <mingyang.sun@arm.com>
Co-authored-by: Ken Liu <ken.liu@arm.com>