Tools: Make manifest tool aware of Partition Enabled status
The manifest tool does not know the enabled status of Partitions.
So it includes every Secure Partition in manifest list even when
some Partitions are disabled.

This patch makes the tool aware of the Partition enabled status basing
on the following feature:
The CMake command configure_file substitutes variable values referenced
as @VAR@ or ${VAR} in the input file content.

By adding "@" to "conditional" attribute in manifest lists, the
manifest tool will get the "conditional" replaced by real enabled status
such as "ON" and "OFF". And skips Partitions that are disabled.

Change-Id: I5ef8a0b5dfb17fd8eb3b5aaec054bc83cdacd8a1
Signed-off-by: Kevin Peng <kevin.peng@arm.com>