Tools: Merge back platform specific manifest list back
As the manifest tools is aware of Partition enabled status, platform
specific manifest list is not needed.
Platforms only need to set correct Partition configurations.

This patch removes the manifest_list_with_se.yaml and adds the proxy
Partition back to default manifest list.
It also add a default configuration of the Proxy Partition as needed
by the manifest tool.

Change-Id: Iee2126b1f038d17291a7e71bf02f4096f59f5ed9
Signed-off-by: Kevin Peng <kevin.peng@arm.com>