SPM: Update boundaries for the 1st partition
The 1st partition's boundary activating is missed, hence it
inherits the initial boundary setting when it gets executed for
the first time. The initial boundary belongs to the PRoT domain,
which is incorrect to be inherited by a partition.
Activate 1st partition's boundary when it is chosen to be run
as the first.

Change-Id: Ia159d62595c15d9716a8271a746761393c772329
Signed-off-by: Ken Liu <Ken.Liu@arm.com>