doc: ps: Small language corrections
The first paragraph first states that hardware isolation backs the
service and then states what happens when hardware isolation is not
present. This is contradictory.

Several paragraphs use language that imply that PS is always used for
external flash when in reality it could be used for internal or
external flash.

The asset size limitation is repeated unnecessarily.

Change-Id: I7853a9e691b37575be59c826bb6029327b43c188
Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>