Tools: Introduce "weak" dependency in Partition manifest
"Weak" dependency is:
A Secure Partition can still work even when the Partitions of the
dependency services are not enabled. This is done by Partitions
checking the Partition enabled status in code.
This kind of dependencies are "weak" dependencies.

The tfm_secure_client_service Partition is the typical one as it only
tests on those services that are enabled.

While "strong" dependency the original definition of dependencies in
FF-M where a Partition cannot work without its dependency.
For example, the Crypto Partition "strongly" depends on ITS services.

"Weak" 'dependencies' list in Partition load info should optionally
includes the SIDs of the dependencies.
Otherwise, build would fail if some Partitions/services are disabled
because the SID is not defined.

This patch:
- Allows a "weak_dependencies" in Partition manifests
- The load info template appends "#ifdef SOME_SID" to the "weak"
  dependency SIDs

Note:
This "weak" dependency is a TF-M specific feature.
It should be only used for test purposes.

Change-Id: Ib0f14f262f8dc4b896666a9558a23d73b24a73d6
Signed-off-by: Kevin Peng <kevin.peng@arm.com>