SPM: Remove unused ER_INITIAL_PSP in IPC model
ER_INITIAL_PSP was used as the ns agent partition stack.
The ns agent changes to use private variable as the stack.
ER_INITIAL_PSP is not used now in IPC model.

This patch removes use of ER_INITIAL_PSP in IPC model isolation
HAL, and in the linker script.

Change-Id: I409f59e19d47e5dbcae08f98d25c47a865bf5200
Signed-off-by: Mingyang Sun <mingyang.sun@arm.com>