Tools: Combine manifest lists and extra manifest lists
The manifest tool takes two manifest lists:
The default manifest list and the extra manifest list for out-of-tree
build partitions.

The processing is actually the same on the two lists.
The difference part is on the paths.

Paths used in the default manifest list are relative paths to TFM root
dir while paths used in extra manifest lists are relative path to the
manifest lists.

So the default manifest list (TFM_MANIFEST_LIST) can be only set to
manifest lists with TF-M repo.

This patch unifies paths usage and then the manifest lists can be
combined together.
With this patch, all manifest lists - whether within TF-M or out of
TF-M - are together pre-processed by CMake and passed to manifest tool.
And the default manifest list can be set to any manifest list.

Change-Id: Id2ae2a4407b1c191f175e7a5ad54edbaecc19523
Signed-off-by: Kevin Peng <kevin.peng@arm.com>