SPM : Arch : Force PendSV triggering under thread mode
Because PendSV deferring is allowed in the architecture,
add 'DSB' and 'ISB' to ensure the PendSV trigger immediately
when the ICSR bit is set under Thread mode, if PendSV has a
higher priority than the current execution. This could ensure
the caller can update the state value into caller context.

In the case that PendSV has a lower priority than the current
execution (Funtion get called in SVC mode e.g.), it is still
workable because the later PendSV execution can also set the
correct return value into the caller's context.

NOTE:
  The debugger MAY disable interrupts during Debug state stepping,
  hence the developer may observe the interrupt does not trigger
  (Even tail-chain case) after it is set by software. Do not step
  when debugging interrupt setting code, set a breakpoint and go
  can be a workaround.

Change-Id: I222b1462266417a1982286c00672d9c6793da043
Signed-off-by: Ken Liu <Ken.Liu@arm.com>