Tools: FF Manifest using "output_path" instead of "source_path"
The manifest list for Firmware Framework test is a "out-of-tree" list.
With the support of out-of-tree build, the special handle on
"source_path" is replaced by "output_path" which gives move flexibility

This patch adds "output_path" to replace "source_path" in manifest list
and removes the handling in manifest tool as well.

Change-Id: I6a975bbc3c847400fd5d1472e281bf8b9e8ae257
Signed-off-by: Kevin Peng <kevin.peng@arm.com>