Platform: Fix main stack run out issue
The size of buffer used in NV counter read/write operation shall
be the same as BLOCK_SIZE which can be 0x800 on NXP platform.
It may run out of main stack if NV counter operation buffer is
located in main stack. This patchm moves the buffer from stack
to bss section.

Signed-off-by: Sherry Zhang <sherry.zhang2@arm.com>
Change-Id: I5c95363a6f52175b9acc2fc8901e926e5b7d5077