Move ns_agent_mailbox partition to manifest scheme
Rather than using hand-coded load_info, add a manifest file for
the new partition like the majority of the partitions.

I added a dummy service because a partition needs to either have
a service or handle an IRQ and we haven't yet got the IRQ code.
The dummy service will be removed when the actual IRQ handling
is added.

This also switches it to use a dedicated stack rather than
ER_INITIAL_PSP. Stack size is carried over.

Change-Id: I6fd2b2ac34cb1ec24755efa72d3e6822817c5c8d
Signed-off-by: Chris Brand <chris.brand@cypress.com>