Tools: Relax the services and irqs list validation
FF-M requires a Secure Partition to have at least one serivces or
irqs (in manifest).
This check is done in template.

This patch moves the check to the manifest tool and perform it only on
Secure Partitions whose PIDs are greater than 256 (TFM_SP_BASE).
PIDs less than 256 are reserved for TF-M special SPs so the check is
not necessary.

Since the audit log Partition is not FF-M compliant, this patch also
change its PID to be less than 256 to skip FF-M compliance check.

Change-Id: I49a13e95e6eb191ae8f2e1bc57190f8d5c7e590c
Signed-off-by: Kevin Peng <kevin.peng@arm.com>