Tools: Allow unset CMake variable in manifest conditions
In CMake, when a variable is not set it is taken as false.
In many cases, Secure Partition build config flag may not be explicitly
set when the SP is not enabled. It makes that config flag as undefined.
The manifest tool would report error on these cases.

This patch adds an empty string in valid disabled conditions for
manifests to adapt more use cases.

Change-Id: I097c21caeb725ea96a093ba625545ff36250aff6
Signed-off-by: Kevin Peng <kevin.peng@arm.com>