platform: lairdconnectivity: BL5340: Add support for platform interrupts
Add platform interrupts for BL5340 for the set of HW
peripherals that the application should be able to use from out-of-tree
secure partitions.

Rename TFM periheral defines to be aligned with naming in nrfx:
 - usbregulator -> usbreg
 - dppic -> dppi
 - pX -> gpioX

Add a tfm_peripherals_config.h header which determines which HW
peripherals can be configured as secure. This saves memory resources
by not including platform_data in mmio list or interrupt handling for
peripherals not needed in the secure image.
This header can be generated from an out-of-tree build and overide the
include header.

The tfm_peripherals_config header also adds support for defining the
GPIO pin and DPPI channel mask. This is needed when these peripherals
are using SPLIT security and dividing the pins and channels among
secure and non-secure.

Remove binding between startup file interrupt vector table names
and nrfx driver interrupts handlers by not including nrfx_irqs.h
In TF-M these handlers should be called either from the second level
interrupt handling by waiting for the {{name}}_SIGNAL signal. Or from
the first level interrupt function {{name}}_flih.

Change-Id: I5224c97145f695d3e0cc4901cea08d95673e17ef
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>