Platform: Refine image flash layout
Eliminate the unnecessary space between the LR_VENEER section
and the rest of the secure code. Place the CMSE veneers right
after the vectors section. The size of vectors section is fixed
for each platform. So the start address of the CMSE veneers
section is also fixed.
Add "a" attribute for the vectors section in startup.S file for
gcc so that the vectors section can be kept in the .bin file.
This patch covers all the platforms that use the tfm_common_s.
sct/ld link script.
Move the .ER_EFLASH_DRIVER_RO section in front of .ER_TFM_CODE
section in gnulinker so that the flash driver code can be picked
into the .ER_EFLASH_DRIVER_RO sectio instead of .ER_TFM_CODE
section.

Signed-off-by: Sherry Zhang <sherry.zhang2@arm.com>
Change-Id: I0084aef576d90af45513c7eaae58ba6c6dbbddab