platform: nordic_nrf: Update nordic HAL handling and version
Update the way the nordic HAL is integrated into the nordic platform.
Provide a way for the nordic HAL path to be set by the application or
downloaded from a specific version.
This allows the application to use the same version of the nordic HAL
for both the secure and non-secure image.
The version that is downloaded is set to include nrfx 2.6.0.
This is an upgrade from what was included, which was nrfx 2.4.0.

Change-Id: Ib20be5428645e0aa178945e2c95ce7294fe6c5c4
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
(cherry picked from commit fc3c1f95b08e3b1c70dd34e3099092fd37acdced)