SPM: Trigger scheduler at the end of interrupt handling
The interrupt handling used to trigger scheduler in thrd_wake_up(),
called by spm_assert_signal().
A recent change moves out the triggering from thrd_wake_up() to the
caller.
And the interrupt handling part was missed.
This does not cause functional issues but may cause priority inversion
when:
Handling of an interrupt - whose owner Partition has higher priority -
requires sheduling from a lower priority Partition that was
preempted. (Typically a SLIH interrupt could have such cases.)
This is because the scheduler was not triggered and the interrupt
handler returns back to the preempted Secure Partition.

This patch adds scheduler triggering back to interrupt handling.

Change-Id: I7ea1af91b89214e6ab401a9ab5ba68384e275c20
Signed-off-by: Kevin Peng <kevin.peng@arm.com>