SPM: Rename 'thread call' with 'cross call'
Name "thread call" could not explicitly reflect the behaviour
of call type. This call crosses between caller and callee stack,
which enables caller and callee context to be sealed in their stack.

Rename this call with "cross call" to make the code more readable.
It will also benefit the upcoming optimization work.

Signed-off-by: Xinyu Zhang <xinyu.zhang@arm.com>
Change-Id: I7d2de563b6f498bde70e302065e743344d2a20fc