SPM: Turn bi-directional list into uni-directional
Uni-list works well enough for linking items and easier to apply
synchronization mechanisms for it as it has only one member to
be maintained when inserting/removing.

The existing usage defines a a fixed header node to make the general
uni-directional list work easier. This fixed header node does not
need to be the same type as the list item, just need to have the
same for link indicator. With this fixed header, there is no need
to keep checking if the head is going to be updated when updating
the members of the list.

For example, a 'partition_t' as the head node and it contains the
link indicator 'p_messages' to link all pending messages. So does
the list usage in handle pool logic.

And it is still applicable to reference these uni-directional list
functions for the general list usages without a fixed head member,
just choose proper functions to match the purpose.

Change-Id: I38a04f59e5fbf799a449a38fa09ea103b554c6bd
Signed-off-by: Ken Liu <Ken.Liu@arm.com>