Platform: Move data_width_byte variable from RAM to Flash
If the const variable is defined as local variable, then it
is stored in Flash and loaded to RAM(stack) when it runs.
If it is defined as file static, its runtime address will be
allocated in Flash. So copy to RAM is avoided and Flash cost
remains the same.

Signed-off-by: Sherry Zhang <sherry.zhang2@arm.com>
Change-Id: I701a6cf01abc5a6d587aa7c263dec0a221487e5c