SPM: Use flag to indicate NS Agent
NS Agent is recognized by partition ID in the existing implementation,
this needs the implementation to reserve a specific value for it
which is irrational for the significance of partition ID.

Assign a flag bit to indicate if a partition is one NS Agent or not.

This also helps abstract the coming memory-checking functionality -
the platform can perform the checking by given partition's boundaries
instead of leaving those platform-specific settings to SPM common
logic, such as executable, read-only, privileged or non-secure
accessible.

Change-Id: I06c56077afc38b1cd2eed53a226c46823896ecd5
Signed-off-by: Ken Liu <Ken.Liu@arm.com>