Proxy: Add a dummy connection-based service
If there is not any connection-based service being built, the
psa_set_rhandle API will be excluded in building as it's for
connection-based serivce only.

But the Proxy Partition requires this API to save handles.
So to avoid the psa_set_rhandle API beening excluded, a dummy
connection-based services is added.

Change-Id: I965bf90c6b47bfff1a52ffbf66ca24d567446801
Signed-off-by: Kevin Peng <kevin.peng@arm.com>