BL2: Add pre and post image load steps to the HAL
The boot_platform_pre_load() and boot_platform_post_load() functions
provide hooks for platforms to add platform-defined steps that need to
be taken before or after a particular image is loaded.

Also, load the images one at a time using boot_go_for_image_id so that
the pre and post steps can be called.

Change-Id: Ic84368754de29bbf2817dd125e7c09a95b947c5f
Signed-off-by: Raef Coles <raef.coles@arm.com>