Tools: Make verify use verifier hierarchy
Change the way the verification runs, so that the process expects a
"hierarchy" of verifiers. This points into the direction that new token
formats could be assembled from basic verifier building blocks.

This would allow to reuse existing code more easily, and extend the
number of token formats recognised by the tool more easily.

Change-Id: Ia4a2b7f8442f1f47e3e613cd1fc68c20e7b74b5a
Signed-off-by: Mate Toth-Pal <mate.toth-pal@arm.com>