SPM: Scheduler: Check sp_limit by CURRENT_THREAD
In scheduler, there is a check to see if there are enough space
on stack to push contexts, to prevent stack overflow.
But the wrong context pointer is used - The one in partition struct
is used.

In fact, the context of the partition struct does not always present
the current context.
Instead, the one in CURRENT_THREAD does.

This patch fixes this issue.

Change-Id: I4c018625395927d396297046ebb90b7eae806bd0
Signed-off-by: Kevin Peng <kevin.peng@arm.com>