platform: nordic: Add option for detecting null-pointer dereferences
A common bug is to dereference a pointer that is NULL, which will
result in a read to address 0. Or to dereference a null-pointer to a
struct, which will result in a read to an address that is smaller than
the size of the struct, often less than, say 256 bytes.

When this happens in the NSPE it results in a SecureFault and no harm
is done. But when this happens in the SPE the read will be sucessful
and program execution will continue with a corrupted value.

This corrupted program could do anything, and will usually result in
an obscure fault that is difficult to debug and possibly even a
security issue.

To prevent this we introduce the platform-dependent opt-in option
NULL_POINTER_EXCEPTION_DETECTION.

NULL_POINTER_EXCEPTION_DETECTION uses two armv8m MPU memory regions to
block all access (except for exception vector fetches) to the first
256 bytes of flash.

It can be enabled either to debug faults in the SPE or as added
security if enough memory regions are available.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>
Change-Id: I5dc65cbc99103fc78dd0f64c07bc5ef82ccd64ab