Tools: Support for dual-model Test Secure Partitions
A TF-M regression test Partition may support both the SFN and IPC model.
The actual model being used follows the SPM backend enabled.
This is quite useful if a Secure Partition is required in testing both
the IPC and SFN backend.

This patch adds this support by allowing Secure Partitions to set
the "model" attribute to "dual".
Then the manifest tool changes it to the corresponding values according
to the selected backend.

Change-Id: I56225b10d3f2384a99e19be3fefaa0e52685a8b0
Signed-off-by: Kevin Peng <kevin.peng@arm.com>