SPM: Introduce option for halting instead of rebooting on panic
Introduce the option to halt instead of reboot to retain the backtrace
that triggered the fault and thereby make it easier to debug.

In addition to retaining the backtrace it is also easier to debug a
program that has halted than one that infinitely reboots.

Also, introduce tfm_hal_system_halt() to allow platform specific
behaviour when halting the system.

Change-Id: I965bf90c6b47bfff1a52ffbf66ca24d567456801
Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>