build: Make ARMClang maximum version check a warning
The maximum allowed version of ARMClang is capped at 6.14 due to a
potential MemManage fault observed with newer versions of the toolchain.
But before the issue becomes well-understood with a fix or workaround
available, we should generate a warning instead of a fatal error and
give users a choice to proceed with a newer version of ARMClang if
they need to (e.g. in order to support some newer hardware platforms).

Change-Id: I3bfcb6b7ae229886d7492fc22b8c40a2e31d54a5
Signed-off-by: Lingkai Dong <lingkai.dong@arm.com>