Explicitly search for the C++ cross compiler
Although TFM is written entirely, it is (almost) possible to write
trusted applications in C++.  However, the TFM build script, at least
for the gnuarm toolchain, only explicitly sets the C compiler not the
C++ compiler.

Without this set, Cmake tries to guess the C++ compiler, and depending
on the build environment seems to either choose the correct cross
compiler, or chooses the host C++ compiler, which results in confusing
compilation failures.

Fix this by explicitly setting the C++ compiler.  This does not add a
test for this, because although most installs of armgcc will have a C++
compiler available, not everyone will be building with it, and it isn't
strictly required.

Signed-off-by: David Brown <david.brown@linaro.org>
Change-Id: Ia1b824dbbb12deb1cb153d84132533d13c417638