CC3XX: Fix importing RSA keys in PSA driver
The CryptoCell representation of RSA keys parameters
does not use the DER encoding as the PSA APIs are
using. The DER encoding adds a leading zero when the
MSb of a number is set. This fix removes the leading
zero from the RSA parameters before they are used
by the CyrptoCell functions.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Change-Id: Id55c1718764ad8c85a1788c243606047c96fe19c