Attest: Store the token buffer in data segment
The local buffer for the created token has a platform dependent size.
Make the token buffer global (thus, it will be stored on the data
segment) avoiding the need to greatly increase the partition's stack
size so it can accomodate the largest tokens.
Decrease the size of stack with the (smallest) size of the buffer.

Signed-off-by: David Vincze <david.vincze@arm.com>
Change-Id: Ic9dfa20de054da4deba74142dfabc1734a6ccdcb