CC3XX: Define a Chacha20-Poly1305 state context
The Chacha20-Poly1305 algorithm implementation for multipart
needs to have an associated context to keep track of the Poly1305
state across multipart API calls to both update() and update_ad().
Thus far single part and Chacha20 multipart were using the existing
ChachaContext_t but that needs to be extended for Chacha20-Poly1305.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: I37d71da465a81754a57d4637f87cc849dd81c33f