CC312: Refactor Poly1305 APIs to support multipart flows
Currently the Poly1305 module in the CC-312 low level stack
supports only integrated operations for Poly1305 MAC computation,
i.e. it will compute the MAC of a message with a single call.
This patch refactors the module to expose APIs callable from a
multipart call flow from a driver interface.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: Ib89e1d8ce03e0592de7fa91171b3d027cb20b11e