Platform: Polaris: Align addresses with FVP
Updated base addresses, the interrupt map, and memory
map to align it with the FVP.

The addresses of INITSVTOR and INITNSVTOR were changed
in the FVP, so the images have to be loaded to different
addresses and bl2 has to be placed in the SRAM, instead
of the ITCM.

The Secure DATA is moved to the first half of ISRAM0 and
the Nonsecure Data is moved to the second half of
ISRAM0. They occupy 1-1 MB of memory each, and ISRAM1
is available for future use.

Change-Id: I82d3c3be69855472a28b8c97885c67b634d2d412
Signed-off-by: Bence Balogh <bence.balogh@arm.com>